feat: BLE_LED_TRIPLE mode, batt_period_s uint16, charts UX
Firmware (TX_Beacon):
- app_ble: add BLE_LED_TRIPLE (mode 2) — 3 quick blinks at adv start,
  then CPU1 uses WFI instead of UartCmd_Wait (fixes SysTick overcount
  in Stop1 that caused session timeout to fire 20x too fast)
- main.c: split BLE sleep path per mode — NORMAL keeps UartCmd_Wait,
  OFF/TRIPLE use WFI loop so BLE IPCC events are still serviced
- flash_config: extend batt_period_s to uint16 via new batt_period_s_ext
  field (backward-compatible); was capped at 255s, now supports up to 3600s
- drv_batt_adc: g_batt_period_s uint8→uint16; default changed to
  PERIODIC/3600s (was OFF/5s)
- drv_chip_temp: add __DSB() before ADC clock disable to prevent
  VREFEN/TSEN write being lost in Cortex-M4 write buffer (~11 µA fix)
- cmd_layer: batt interval range 1-3600s (was 1-255); config reset
  defaults to PERIODIC/3600s
- svc_uart_cmd: batt period range 1-3600 s
- proto_uart: remove batt_period_s uint8 cast
- cmd_layer: accept ble_led_mode up to 2 (TRIPLE)

Flutter app:
- home_tab: value cards (temp/battery/light) are now tappable,
  open live chart in a modal bottom sheet
- data_tab: fullscreen button on each chart card (push new route)
- logging_tab: add Ukrainian time-depth formatter (_fmtDepth)
- beacon_provider: _statusRefreshing guard prevents race between
  background poll and writeConfig; suppressActivity during readStatus
- ble_settings: ledMode comment updated to reflect 3 modes

PC app (tx_beacon_gui_v4.py):
- BLE LED mode: replace checkbox with 3-item ComboBox
  (Normal / Off / Triple blink)

+++ TX_Beacon/Core/Inc/drv_batt_adc.h
@@ -18,8 +18,8 @@
 #define BATT_MODE_OFF       0U
 #define BATT_MODE_PERIODIC  1U
 
-extern uint8_t g_batt_mode;
-extern uint8_t g_batt_period_s;
+extern uint8_t  g_batt_mode;
+extern uint16_t g_batt_period_s;
 extern uint8_t g_batt_scale_x10;    /* divider multiplier ×10 (20 = 2.0×) */
 
 extern uint32_t g_last_batt_mV;
